move organisation graphics and info into /app/about/_components.
add way of enabling group of circles by opacity.

edit info text
