+++ src/app/about/_components/OrganisationMembersGraphic.tsx
@@ -1,12 +1,12 @@
 "use client"
-import { OrganisationMembersInfo } from "@/app/_components/OrganisationMemberInfo"
+import { OrganisationMembersInfo } from "@/app/about/_components/OrganisationMemberInfo"
 import { CircleDashed } from "lucide-react"
 import { useEffect, useState } from "react"
 
 /*
 POSSIBLE IMPROVEMENTS:
 amount of layers change dynamically
-...
+better way of disabling circles - currently by opacity
 */
 
 export function OrganisationMembersGraphic() {
@@ -18,6 +18,7 @@ export function OrganisationMembersGraphic() {
 
     let pi_division = 0
     let pi_division2 = 0
+    const opacity_constant_pg = 1
     const pg_circles = []
     const pg_layer1 = Math.floor(
       Math.PI / Math.asin(memberRadius / layerRadius)
@@ -36,7 +37,8 @@ export function OrganisationMembersGraphic() {
         name: i,
         cx: 85 + layerRadius * Math.sin(i * pi_division + theta),
         cy: 85 + layerRadius * Math.cos(i * pi_division + theta),
-        r: memberRadius
+        r: memberRadius,
+        opacity: opacity_constant_pg
       })
     }
     for (let i = pg_layer1; i < people; i++) {
@@ -44,7 +46,8 @@ export function OrganisationMembersGraphic() {
         name: i,
         cx: 85 + 2 * layerRadius * Math.sin(i * pi_division2 - theta),
         cy: 85 + 2 * layerRadius * Math.cos(i * pi_division2 - theta),
-        r: memberRadius
+        r: memberRadius,
+        opacity: opacity_constant_pg
       })
     }
     //animating the circles
@@ -67,6 +70,7 @@ export function OrganisationMembersGraphic() {
     let pi_division = 0
     let pi_division2 = 0
     let pi_division3 = 0
+    const opacity_constant_ot = 1
     const ot_circles = []
     //amount of circles of radius [memberRadius] that can fit on the circle making up a layer
     const ot_layer1 = Math.floor(
@@ -94,7 +98,8 @@ export function OrganisationMembersGraphic() {
         name: i,
         cx: 85 + 3 * layerRadius * Math.sin(i * pi_division + theta),
         cy: 85 + 3 * layerRadius * Math.cos(i * pi_division + theta),
-        r: memberRadius
+        r: memberRadius,
+        opacity: opacity_constant_ot
       })
     }
     for (let i = ot_layer1; i < ot_layer1 + ot_layer2; i++) {
@@ -102,7 +107,8 @@ export function OrganisationMembersGraphic() {
         name: i,
         cx: 85 + 4 * layerRadius * Math.sin(i * pi_division2 - theta),
         cy: 85 + 4 * layerRadius * Math.cos(i * pi_division2 - theta),
-        r: memberRadius
+        r: memberRadius,
+        opacity: opacity_constant_ot
       })
     }
     for (let i = ot_layer1 + ot_layer2; i < people; i++) {
@@ -110,7 +116,8 @@ export function OrganisationMembersGraphic() {
         name: i,
         cx: 85 + 5 * layerRadius * Math.sin(i * pi_division3 + theta),
         cy: 85 + 5 * layerRadius * Math.cos(i * pi_division3 + theta),
-        r: memberRadius
+        r: memberRadius,
+        opacity: opacity_constant_ot
       })
     }
     //animating the circles
@@ -133,6 +140,7 @@ export function OrganisationMembersGraphic() {
     let pi_division = 0
     let pi_division2 = 0
     let pi_division3 = 0
+    const opacity_constant_host = 1
     const host_circles = []
     //amount of circles of radius [memberRadius] that can fit on the circle making up a layer
     const host_layer1 = Math.floor(
@@ -160,7 +168,8 @@ export function OrganisationMembersGraphic() {
         name: i,
         cx: 85 + 6 * layerRadius * Math.sin(i * pi_division - theta),
         cy: 85 + 6 * layerRadius * Math.cos(i * pi_division - theta),
-        r: memberRadius
+        r: memberRadius,
+        opacity: opacity_constant_host
       })
     }
     for (let i = host_layer1; i < host_layer1 + host_layer2; i++) {
@@ -168,7 +177,8 @@ export function OrganisationMembersGraphic() {
         name: i,
         cx: 85 + 7 * layerRadius * Math.sin(i * pi_division2 + theta),
         cy: 85 + 7 * layerRadius * Math.cos(i * pi_division2 + theta),
-        r: memberRadius
+        r: memberRadius,
+        opacity: opacity_constant_host
       })
     }
     for (let i = host_layer1 + host_layer2; i < people; i++) {
@@ -176,7 +186,8 @@ export function OrganisationMembersGraphic() {
         name: i,
         cx: 85 + 8 * layerRadius * Math.sin(i * pi_division3 - theta),
         cy: 85 + 8 * layerRadius * Math.cos(i * pi_division3 - theta),
-        r: memberRadius
+        r: memberRadius,
+        opacity: opacity_constant_host
       })
     }
     //animating the circles
@@ -201,9 +212,11 @@ export function OrganisationMembersGraphic() {
         cx={circle.cx}
         cy={circle.cy}
         r={circle.r}
+        opacity={circle.opacity}
       />
     )
   })
+  console.log(pgCircles[0].props.opacity)
 
   const otCircles = OtCircleList(80).map(circle => {
     return (
@@ -213,6 +226,7 @@ export function OrganisationMembersGraphic() {
         cx={circle.cx}
         cy={circle.cy}
         r={circle.r}
+        opacity={circle.opacity}
       />
     )
   })
@@ -225,6 +239,7 @@ export function OrganisationMembersGraphic() {
         cx={circle.cx}
         cy={circle.cy}
         r={circle.r}
+        opacity={circle.opacity}
       />
     )
   })
@@ -232,14 +247,15 @@ export function OrganisationMembersGraphic() {
   return (
     <figure className="relative mt-5 w-9/12 object-scale-down">
       <svg viewBox="0 0 170 170" xmlns="http://www.w3.org/2000/svg">
-        <circle fill="orange" cx="85" cy="85" r={memberRadius} />
+        <circle fill="orange" cx="85" cy="85" r={memberRadius} opacity={1} />
         {pgCircles}
         {otCircles}
         {hostCircles}
       </svg>
       <figcaption className="bottom-0 opacity-90 md:absolute">
         <OrganisationMembersInfo title="Infographics guide">
-          The graphics below contain our current organisation structure.
+          Each year, more than 200 students join Armada to make the fair happen!
+          Every circle in the graphic represents one of our members.
           <ul>
             <li>
               <CircleDashed
